dev 2.0

+++ src/main/java/com/fedorov/storage/utils/Delete.java
@@ -1,33 +1,47 @@
 package com.fedorov.storage.utils;
 
+import com.fedorov.storage.repo.FileRepo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.io.File;
 
+@Component
 public class Delete {
 
-    public boolean deleteDirectory(String path, String name) {
+    @Autowired
+    private FileRepo fileRepo;
+    @Transactional
+    public void deleteDirectory(File file) {
 
-        File file = new File(path + "/" + name);
+        String path = file.getPath().replace("\\", "/").substring(0, file.getPath().replace("\\", "/").lastIndexOf("/"));
+        String name = file.getName();
+        System.out.println(path);
+        System.out.println(name);
+
+        if (!file.exists()) {
+            return;
+        }
         if (file.isDirectory()) {
 
+
             File[] files = file.listFiles();
+            if (files != null) {
+                for (File f : files) {
 
-            for (File i : files) {
-
-                deleteDirectory(i.getParent(), i.getName());
+                    deleteDirectory(f);
+                }
             }
         }
 
-        return file.delete();
+        file.delete();
+        fileRepo.deleteByNameAndPath(name, path);
+
 
 
     }
 
-    public boolean deleteFile(String path, String name) {
 
-        File file = new File(path + "/" + name);
-
-
-        return file.delete();
-    }
 
 }
